Fix risk cost migrations and add validation
Add validation to check cost and risk_cost are positive
Stop recalculation of risk_cost
Set risk_cost as integer data type
Fix data migration to update old records

+++ db/migrate/20220707092226_add_risk_cost_to_cost_log.rb
@@ -1,5 +1,5 @@
 class AddRiskCostToCostLog < ActiveRecord::Migration[6.0]
   def change
-    add_column :cost_logs, :risk_cost, :decimal
+    add_column :cost_logs, :risk_cost, :integer
   end
 end
